Location, Camera, Firebase
When user clicks on the New Post button, they take a picture with their phone.

After taking a picture, the user can see the picture they took and they type in the quantity.

When they submit the post, the current date, imageURL of the photo, quantity and location (latitude and longitude) of the user are entered into the Firebase database called 'posts'.

When the user goes back to the homepage (List Screen), they can see the date and quantity of each post in the database.